Renamed builder to webpack for clarity.
